split declaration and implementation
